requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic~=2.10.4
vertexai~=1.71.1
google-genai